Release 0.8.11

+++ evm-contracts/src/v0.6/EACAggregatorProxy.sol
@@ -1,6 +1,6 @@
 pragma solidity 0.6.6;
 
-import "./AggregatorProxy.sol";
+import "./dev/AggregatorProxy.sol";
 import "./interfaces/AccessControllerInterface.sol";
 
 /**
@@ -53,8 +53,9 @@ contract EACAggregatorProxy is AggregatorProxy {
   }
 
   /**
-   * @notice Reads the last updated height from aggregator delegated to.
-   * @dev overridden function to add the checkAccess() modifier
+   * @notice get the latest completed round where the answer was updated. This
+   * ID includes the proxy's phase, to make sure round IDs increase even when
+   * switching to a newly deployed aggregator.
    * @dev deprecated. Use latestRoundData instead.
    */
   function latestTimestamp()
@@ -124,7 +125,9 @@ contract EACAggregatorProxy is AggregatorProxy {
    * data from and validate that they can properly handle return data from all
    * of them.
    * @param _roundId the round ID to retrieve the round data for
-   * @return roundId is the round ID for which data was retrieved
+   * @return roundId is the round ID from the aggregator for which the data was
+   * retrieved combined with a phase to ensure that round IDs get larger as
+   * time moves forward.
    * @return answer is the answer for the given round
    * @return startedAt is the timestamp when the round was started.
    * (Only some AggregatorV3Interface implementations return meaningful values)
@@ -135,17 +138,17 @@ contract EACAggregatorProxy is AggregatorProxy {
    * (Only some AggregatorV3Interface implementations return meaningful values)
    * @dev Note that answer and updatedAt may change between queries.
    */
-  function getRoundData(uint256 _roundId)
+  function getRoundData(uint80 _roundId)
     public
     view
     checkAccess()
     override
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     )
   {
     return super.getRoundData(_roundId);
@@ -160,7 +163,9 @@ contract EACAggregatorProxy is AggregatorProxy {
    * should determine what implementations they expect to receive
    * data from and validate that they can properly handle return data from all
    * of them.
-   * @return roundId is the round ID for which data was retrieved
+   * @return roundId is the round ID from the aggregator for which the data was
+   * retrieved combined with a phase to ensure that round IDs get larger as
+   * time moves forward.
    * @return answer is the answer for the given round
    * @return startedAt is the timestamp when the round was started.
    * (Only some AggregatorV3Interface implementations return meaningful values)
@@ -177,11 +182,11 @@ contract EACAggregatorProxy is AggregatorProxy {
     checkAccess()
     override
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     )
   {
     return super.latestRoundData();
@@ -199,18 +204,18 @@ contract EACAggregatorProxy is AggregatorProxy {
    * @return answeredInRound is the round ID of the round in which the answer
    * was computed.
   */
-  function proposedGetRoundData(uint256 _roundId)
+  function proposedGetRoundData(uint80 _roundId)
     public
     view
     checkAccess()
     hasProposal()
     override
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     )
   {
     return super.proposedGetRoundData(_roundId);
@@ -234,18 +239,19 @@ contract EACAggregatorProxy is AggregatorProxy {
     hasProposal()
     override
     returns (
-      uint256 roundId,
+      uint80 roundId,
       int256 answer,
       uint256 startedAt,
       uint256 updatedAt,
-      uint256 answeredInRound
+      uint80 answeredInRound
     )
   {
     return super.proposedLatestRoundData();
   }
 
   /**
-   * @dev reverts if the caller does not have access by the accessController contract
+   * @dev reverts if the caller does not have access by the accessController
+   * contract or is the contract itself.
    */
   modifier checkAccess() {
     require(accessController.hasAccess(msg.sender, msg.data), "No access");
